fix reported windows errors

+++ batchmerkleproof.go
@@ -39,7 +39,7 @@ func NewBatchMerkleProof(json string) (BatchMerkleProof, error) {
 func (b *batchMerkleProof) Valid(expectedRoot []byte) bool {
 	byteData := C.CBytes(expectedRoot)
 	defer C.free(unsafe.Pointer(byteData))
-	res := C.ergo_lib_batch_merkle_proof_valid(b.p, (*C.uchar)(byteData), C.ulong(len(expectedRoot)))
+	res := C.ergo_lib_batch_merkle_proof_valid(b.p, (*C.uchar)(byteData), C.uintptr_t(len(expectedRoot)))
 	return bool(res)
 }
